# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Generar Token de Acceso (Login)",
                "description": "Autentica al usuario con correo, contraseña y dispositivo para obtener un token de acceso (Sanctum).",
                "operationId": "09f5a11774e5690ff8800e9f498401a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "correo",
                                    "contraseña",
                                    "dispositivo"
                                ],
                                "properties": {
                                    "correo": {
                                        "description": "Correo electrónico del usuario.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@pimos.com"
                                    },
                                    "contraseña": {
                                        "description": "Contraseña del usuario.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "dispositivo": {
                                        "description": "Nombre del dispositivo (Sanctum).",
                                        "type": "string",
                                        "example": "windows"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso. Retorna un token de acceso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inicio de sesión exitoso."
                                        },
                                        "token": {
                                            "description": "Token de acceso (Bearer Token).",
                                            "type": "string",
                                            "example": "lkYTYtksh64dJh9bqKrFa5GHpQZax2IA9D8fdI7Rc54a30ac"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales no válidas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciales inválidas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión y revocar Token",
                "description": "Invalida el token de acceso actual del usuario autenticado.",
                "operationId": "cc362ce60174307c378af272a6efe0ef",
                "responses": {
                    "200": {
                        "description": "Cierre de sesión exitoso. Token revocado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cierre de sesión exitoso. Token revocado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/inventario": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Consultar todo el inventario",
                "description": "Retorna una lista de todos los registros de inventario con los productos asociados.",
                "operationId": "4addcd7798fe1538104ed903e872178f",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InventarioResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/inventario/{producto_id}": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Consultar stock de un producto",
                "description": "Retorna el registro de inventario (stock) de un producto específico por su ID.",
                "operationId": "4872e6af08987b106c16fdfc2ded2b65",
                "parameters": [
                    {
                        "name": "producto_id",
                        "in": "path",
                        "description": "ID del producto para consultar su inventario.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventarioResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Registro no encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Actualizar o Crear Stock",
                "description": "Actualiza la cantidad de existencias para un producto o crea el registro de inventario si no existe.",
                "operationId": "cc8e83f6aa6287050dd4471c3460359f",
                "parameters": [
                    {
                        "name": "producto_id",
                        "in": "path",
                        "description": "ID del producto a actualizar o crear stock.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cantidad_existencias"
                                ],
                                "properties": {
                                    "cantidad_existencias": {
                                        "description": "Nueva cantidad de existencias.",
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock actualizado con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventarioResource"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Inventario inicial creado con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InventarioResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto asociado no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Eliminar registro de inventario",
                "description": "Elimina el registro de stock (inventario) para un producto específico por su ID.",
                "operationId": "c7075e23314e020fdc269fc982bc54a4",
                "parameters": [
                    {
                        "name": "producto_id",
                        "in": "path",
                        "description": "ID del producto cuyo registro de inventario se desea eliminar.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registro de inventario eliminado con éxito."
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Registro no encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InventarioResource": {
                "title": "Inventario Resource",
                "description": "Estructura de la respuesta para un registro de Inventario",
                "properties": {
                    "id": {
                        "description": "ID del registro de inventario",
                        "type": "integer",
                        "example": 5
                    },
                    "tipo": {
                        "type": "string",
                        "example": "inventario"
                    },
                    "atributos": {
                        "properties": {
                            "cantidad_existencias": {
                                "type": "integer",
                                "example": 100
                            }
                        },
                        "type": "object"
                    },
                    "relaciones": {
                        "properties": {
                            "producto": {
                                "description": "Detalles básicos del producto",
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Gomita de Fresa"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Ingrese el token de autenticación con el prefijo 'Bearer '. Ejemplo: Bearer 1|xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}